requirement.txt
# Camera & GPIO
RPi.GPIO
picamera2

# Face Recognition
face_recognition
opencv-python
numpy

# Data Handling & Time
pandas

# PostgreSQL
psycopg2-binary

# LCD Display
RPLCD

# Image Processing
Pillow

# HTTP Requests
requests

# Scheduling
APScheduler
